+++ packages/lib/src/model/BaseModel.ts
@@ -129,6 +129,7 @@ export abstract class BaseModel<
     }: ModelConstructorOptions = arguments[1] as any
 
     Object.setPrototypeOf(this, modelClass!.prototype)
+    this.constructor = modelClass!
 
     const self = this as any